# [Core Crypto] Optimize hash