# Description

This pr optimize the hash algorithm by using static method, quote:

> CA1850: Prefer static HashData method over ComputeHash
> It's more efficient to use these static HashData methods than to create and manage a HashAlgorithm instance to call ComputeHash.


bench with ` var data = Encoding.ASCII.GetBytes(string.Concat(Enumerable.Repeat("Hello, World!^_^", 16)));`

| Method                | Mean       | Error     | StdDev    |
|---------------------- |-----------:|----------:|----------:|
| Sha256_ComputeHash    | 355.224 ns | 1.7175 ns | 1.6065 ns |
| Sha256_HashData       | 227.764 ns | 1.2923 ns | 1.1456 ns |

bench with  ` var data = Encoding.ASCII.GetBytes(string.Concat(Enumerable.Repeat("Hello, World!^_^", 1)));`

| Method                | Mean        | Error     | StdDev    |
|---------------------- |------------:|----------:|----------:|
| Sha256_ComputeHash    | 250.2674 ns | 1.9593 ns | 1.8327 ns |
| Sha256_HashData       | 126.8578 ns | 0.4144 ns | 0.3674 ns |



Fixes # 

## Type of change

<!-- Please delete options that are not relevant. -->

- [x] Optimization (the change is only an optimization)
- [ ] Style (the change is only a code style for better maintenance or standard purpose)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update


# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules