@Jim8y 2024-12-30
> Which platform did you test it on?

windows 11 intel i9-14900HX